requirements.txt
Flask
google-generativeai
youtube-transcript-api
python-dotenv
